+++ example/lib/success.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+
+class Success extends StatefulWidget {
+  const Success({Key? key}) : super(key: key);
+
+  @override
+  State<Success> createState() => _SuccessState();
+}
+
+class _SuccessState extends State<Success> {
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: MediaQuery.of(context).size.width,
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            const SizedBox(
+              height: 10,
+            ),
+            const Center(
+              child: Text(
+                'Failure',
+                style: TextStyle(color: Colors.redAccent, fontSize: 40, fontWeight: FontWeight.w600),
+              ),
+            ),
+            const SizedBox(
+              height: 25,
+            ),
+            const Text(
+              'Your application has been rejected. Please fill the form properly and upload again.',
+              style: TextStyle(color: Colors.black87, fontSize: 22, fontWeight: FontWeight.w400),
+            ),
+            const SizedBox(
+              height: 25,
+            ),
+            const SizedBox(
+              height: 20,
+            ),
+            ClipRRect(
+                borderRadius: BorderRadius.circular(10),
+                child: Container(
+                  color: Colors.yellow,
+                  height: 50,
+                  child: const Center(
+                      child: Text(
+                    'OKAY',
+                    style: TextStyle(fontSize: 18, fontWeight: FontWeight.w600),
+                  )),
+                ))
+          ],
+        ),
+      ),
+    );
+  }
+}