+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## 0.0.3
+
+[0.0.2] - 07.02.2023
+
+Added iconBackgroundColor, childBackgroundColor and position for iconWidget.
+
+## 0.0.2
+
+[0.0.2] - 06.02.2023
+
+Updated documentation and added new images
+
 ## 0.0.1
 
 [0.0.1] - 06.02.2023
 
-Initial publish of project containing IconBottomSheet
-Demo app using icon_bottomsheet package
+Initial publish of project containing IconBottomSheet Demo app using icon_bottomsheet package